pyproject.toml
[project]
name = "chatbot-metadata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.1",
    "bcrypt>=4.2.1",
    "colorlog>=6.9.0",
    "elasticsearch>=8.17.1",
    "fastapi>=0.115.8",
    "openai>=1.61.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "sqlalchemy>=2.0.37",
    "uvicorn[standard]>=0.34.0",
]
